lwm2m: FOTA updates for AT&T
A lot of fixes.

- Refactored Firmware object callback
- All readable Firmware resource are now discoverable
- Resources /5/0/3 and /5/0/5 are now observable
- Added support for whole-object TLV writes to Firmware object
- Added callback on TLV writes to Firmware object to trigger FOTA
- Added parsing of port in package URI
- Added size check on firmware image, needed by AT&T for ATT16
- Report `Connection lost` when an error (presumably a socket error)
  occurs during download, needed by AT&T for ATT17
- Improved general handling of errors during download
- Fixed a bug where `Update Result` resource initial state
  was not set when FOTA was started, failing back-to-back tests
- Correctly retrieve APN based on carrier during FOTA
- Fixed a bug where the download task was restarted after requesting
  to apply a modem patch, when failing to parse the HTTP header
- Added TLS certificate for AT&T
- Added abstraction for lwm2m_os_key_mgmt_cmp() to compare certs
- Improved certificate handling to detect whether we need to
  reprovision a certificate upon boot
- Added LWM2M_OBJECT_INSTANCE define
- Ninja fixes to Server object encoding
- Pulled in latest NCS changes
- Increased AT_CMD_RESPONSE_MAX_LEN by 100 bytes to accommodate
  reading out the certificates already provisioned
- Minor fixes to remove compilation warnings in newer GCC

+++ client/src/lwm2m_shell.c
@@ -596,7 +596,7 @@ static void dump_as_hex(const struct shell *shell, char *data, int len)
 
         // Dump bytes as text
         for (i = 0; i < 16 && pos + i < len; i++) {
-            if (isprint(data[pos + i])) {
+            if (isprint((int)data[pos + i])) {
                 *cur++ = data[pos + i];
             } else {
                 *cur++ = '.';
